Debug exercises

+++ lab2/DebugPractice/DebugExercise1.java
@@ -3,8 +3,8 @@
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+    public static int divideThenRound(float top, float bottom) { // change the type of params
+        float quotient = top / bottom;
         int result = Math.round(quotient);
         return result;
     }
